Piscine : une fonction = une page (activation, relais, parametres)
pdN et dNN etaient le meme objet sous deux noms, regles dans deux menus
differents : brancher une fonction demandait trois pages sans lien visible
(page Equipements, champ *_slot de la branche metier, io/output/dNN).

Les trois couches restent inchangees -- PoolLogic pour les regulations,
PoolDevice pour les interlocks/quotas/comptages, IOModule pour le binding.
Seule leur presentation change : une fonction = un PoolDevice = la sortie
logique dNN de meme index, et le seul choix laisse a l'utilisateur est le
relais physique.

12 fonctions renumerotees dans PoolIds::Device*, avec ActuatorO2Pump et
ActuatorAux1 nouveaux. La pompe oxygene actif est un appareil distinct du
chlore : debit, bidon et compteurs de consommation separes, pour zero
nouvelle cle NVS (les 16 descripteurs de PoolDeviceSlots.h existaient deja).

  ** Effacement NVS obligatoire ** : pd3* designait le robot et designe
  desormais l'electrolyseur, pd4* passe de RELAY_STD a PERISTALTIC.

Arbre de config piscine/* obtenu sans renommer aucune cle NVS, via les alias
d'affichage et meta.order (implemente dans app.js mais declare nulle part).
Ordre racine declare en entier : ordonner une seule branche suffisait a
deplacer toutes les autres, ce qui reléguait le menu Home Assistant.

Supprime : les 7 champs *_slot et leurs cles pl_s*, l'enum_set
poollogic_device_slot, la page Equipements, ~90 lignes de JS mort et
5 exports I2C par cycle dans HMIModule devenus sans objet.

Corrige trois defauts preexistants rencontres en chemin :
- appliquer une valeur sur une page composee la cassait (toJsonModule fait un
  strcmp exact, le rechargement post-apply partait en 404) ;
- dependsOnMask tronquait en silence au-dela du slot 7 (uint8_t avec des
  casts (uint8_t)(1u << i) dans des boucles allant jusqu'a 16) ;
- requireSetup briquait la carte sur une table mal formee (while(true)) ;
  l'invariant pdN <-> dNN est desormais un static_assert, verifie en
  injectant une incoherence.

IOModule delie et journalise la seconde sortie quand deux reclament le meme
relais : le risque preexistait, rendre le dropdown accessible depuis chaque
page l'augmente.

Capacites Waveshare relevees (sorties 12, switches HA 20, MaxDomainSlots 28).
8 relais physiques restent le seul plafond non relevable : 12 fonctions
declarees, 8 branchables, les trois modes de desinfection etant exclusifs.

Ecart assume : la temporisation de recopie du debit reste flow_copy_delay_s
plutot que pd9/on_delay_s, le snapshot runtime publiant flow_copy et delay_s
vers des binary_sensors HA. on_delay_s est masque sur pd9/pd10.

Verifie : build Waveshare vert (47,2 % flash), validate_i18n 0 erreur, dette
EN 105 -> 104. Non verifie : le comportement sur carte.

+++ src/Modules/PoolDeviceModule/PoolDeviceLifecycle.cpp
@@ -66,7 +66,7 @@ bool PoolDeviceModule::ensureStorage_()
     }
     if (!slots_) slots_ = allocPsramArray_<PoolDeviceSlot>(POOL_DEVICE_MAX);
     if (!cfgEnabledVar_) cfgEnabledVar_ = allocPsramArray_<ConfigVariable<bool,0>>(POOL_DEVICE_MAX);
-    if (!cfgDependsVar_) cfgDependsVar_ = allocPsramArray_<ConfigVariable<uint8_t,0>>(POOL_DEVICE_MAX);
+    if (!cfgDependsVar_) cfgDependsVar_ = allocPsramArray_<ConfigVariable<uint16_t,0>>(POOL_DEVICE_MAX);
     if (!cfgFlowVar_) cfgFlowVar_ = allocPsramArray_<ConfigVariable<float,0>>(POOL_DEVICE_MAX);
     if (!cfgTankCapVar_) cfgTankCapVar_ = allocPsramArray_<ConfigVariable<float,0>>(POOL_DEVICE_MAX);
     if (!cfgTankInitVar_) cfgTankInitVar_ = allocPsramArray_<ConfigVariable<float,0>>(POOL_DEVICE_MAX);
@@ -210,7 +210,7 @@ void PoolDeviceModule::init(ConfigStore& cfg, ServiceRegistry& services)
         cfgDependsVar_[i].nvsKey = slot.dependsKey;
         cfgDependsVar_[i].jsonName = "depends_on_mask";
         cfgDependsVar_[i].moduleName = slot.configModuleName;
-        cfgDependsVar_[i].type = ConfigType::UInt8;
+        cfgDependsVar_[i].type = ConfigType::UInt16;
         cfgDependsVar_[i].value = &s.def.dependsOnMask;
         cfgDependsVar_[i].persistence = ConfigPersistence::Persistent;
         cfgDependsVar_[i].size = 0;
